apps/ai-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
minio==7.2.8
psycopg[binary]==3.2.9
PyMuPDF==1.26.3
python-docx==1.1.2
httpx==0.28.1
python-dotenv==1.0.1
